requirements.txt
selenium==4.10.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
fastapi==0.109.2
uvicorn==0.27.1
pydantic~=2.6.1
python-dotenv~=1.0.1
python-dateutil==2.8.1
pytest==8.0.1
httpx==0.27.0